Added saveasFile method functionality
The program now is able to save the output of the GUI as a text file.
However, it isn't able to copy over the formatting of text.



I believe this is because we're using StyledDocument rather than making
a string. File writing needs a string to write into the file and
StyledDocument doesn't output a string, unless there's a method to do so
I've been trying to search for one, but can't find one. I recommend
switching the way we're currently formatting the output to a more
primitive way of just manually altering strings.