Added functionality to save and recordError methods and minor changes
Additions:

Save method will now save to the file that was saved to by the save as
method. Save method is not intended to create a new file to be saved to.

recordError method will now save just like the save as method, however,
it will save the error output instead of the formatted text output.

Changes:

Made some additions to existing code to allow for actionListener to be
added to recordErrors.

Changed part of the if statement that differentiates between commands
and text to check for a hyphen being input by itself.

Changed some switch case statements to have error handling for when
additional input is being added.

Renamed a few variables and added spacing to make code easier to read
and comprehend.