Minor Changes & debugging
Fixed single and double spacing to work only on lines after it's
declared and fixed it to be able to output as a text file.

Fixed program not properly adjusting to when a new line length is
declared.

Fixed alignments not working for when a line of text is too long and
needs to move part of the text to the next line.

Made a few additional comments to help better understanding of code.